refactor(risk): move position-state helpers into risk engine
